requirements.txt
langchain-core
langchain-community
langchain-google-genai
langchain-groq
langchain-huggingface
langgraph
matplotlib
oauthlib
pydantic
python-dotenv
gradio
requests
pandas
numpy
Pillow
pytesseract
tavily-python
wikipedia
arxiv
duckduckgo-search
datasets
huggingface_hub